+++ spec/dummy/app/models/company.rb
@@ -1,2 +1,3 @@
 class Company < ActiveRecord::Base
+  # Dummy models
 end