+++ spec/support/apartment_reset.rb
@@ -1,0 +1,9 @@
+module Apartment
+  module Test
+    def self.reset
+      Apartment::Database.instance_variable_set :@initialized, nil
+      Apartment.excluded_models = nil
+      Apartment.use_postgres_schemas = nil
+    end
+  end
+end